Split src code into separate files; added more help api when testing with mocked components
