+++ py/client/zen_snapshot_mock.py
@@ -9,8 +9,8 @@ import json
 LOCAL_HTTP_SERVER_URL = "http://localhost:5000/"
 REMOTE_HTTP_SERVER_URL = "http://zendao-tn-1.de.horizenlabs.io:5000/"
 
-#HTTP_SERVER_URL = LOCAL_HTTP_SERVER_URL
-HTTP_SERVER_URL = REMOTE_HTTP_SERVER_URL
+HTTP_SERVER_URL = LOCAL_HTTP_SERVER_URL
+#HTTP_SERVER_URL = REMOTE_HTTP_SERVER_URL
 
 
 CREATE_PROPOSAL_MOCK = {
