+++ py/server/zen_balancer.py
@@ -1,39 +1,17 @@
-import sys
 import threading
+import cryptography
 from flask import Flask, request, json
 
-from core_methods import print_incoming, add_ownership_entry, \
-    print_outgoing, get_mc_address_map, get_chain_tip, store_proposal_data, get_address_balance, get_active_proposal, \
+from py.server.rosetta.rosetta_methods import get_chain_tip, get_address_balance
+from py.server.snapshot.snapshot_methods import add_ownership_entry, \
+    get_mc_address_map, store_proposal_data, get_active_proposal, \
     proposal_dict
 from definitions import mock_nsc, MOCK_MC_ADDRESS_MAP
+from py.server.utils.util_methods import print_incoming, print_outgoing
 
 
 def api_server():
     app = Flask(__name__)
-
-    @app.route('/api/v1/addOwnership', methods=['POST'])
-    def add_ownership():
-        proposal = json.loads(request.data)
-
-        print_incoming("BalancerApiServer", "/api/v1/addOwnership", proposal)
-
-        if mock_nsc:
-            ret = {
-                'result': add_ownership_entry(proposal),
-                'ownerships': MOCK_MC_ADDRESS_MAP,
-            }
-        else:
-            ret = {
-                "error": {
-                    "code": 109,
-                    "description": "Could not add ownership",
-                    "detail": "Method not supported with real native smart contract. Pls set \'mock_nsc=true\'"
-                              " in balancer"
-                }
-            }
-        print_outgoing("BalancerApiServer", "/api/v1/addOwnership", ret)
-
-        return json.dumps(ret)
 
     @app.route('/api/v1/getOwnerships', methods=['POST'])
     def get_ownerships():
@@ -134,12 +112,36 @@ def api_server():
         print_outgoing("BalancerApiServer", "/api/v1/getVotingPower", response)
         return json.dumps(response)
 
-    app.run(host="0.0.0.0", port=5000)
+    # usable only when mocking native smart contract
+    @app.route('/api/v1/addOwnership', methods=['POST'])
+    def add_ownership():
+        proposal = json.loads(request.data)
+
+        print_incoming("BalancerApiServer", "/api/v1/addOwnership", proposal)
+
+        if mock_nsc:
+            ret = {
+                'result': add_ownership_entry(proposal),
+                'ownerships': MOCK_MC_ADDRESS_MAP,
+            }
+        else:
+            ret = {
+                "error": {
+                    "code": 109,
+                    "description": "Could not add ownership",
+                    "detail": "Method not supported with real native smart contract. Pls set \'mock_nsc=true\'"
+                              " in balancer"
+                }
+            }
+        print_outgoing("BalancerApiServer", "/api/v1/addOwnership", ret)
+
+        return json.dumps(ret)
+
+    context = ('local.crt', 'local.key')#certificate and key files
+    app.run(host="0.0.0.0", port=5000, ssl_context='adhoc')
 
 
 if __name__ == '__main__':
-
-    args = sys.argv[1:]
 
     # Start http server
     t = threading.Thread(target=api_server, args=())
